\name{dTBEinf}
\alias{dTBEinf}
\title{Density of TBE infinity}
\description{Computations for the density of \eqn{\text{TBE}_\infty}, 
including direct calculation, saddlepoint approximation, 
Fourier inversion, and modified W-Transformation.}
\usage{
	dTBEinf(y, X, beta, phi, method="saddle", m = NULL)
}
\arguments{
	\item{y}{Vector of responses \eqn{\mbox{\bf y}\in \mathbb{R}^{n}}.}
	\item{X}{The design matrix \eqn{\mbox{\bf X}\in \mathbb{R}^{n\times p}}.}
	\item{beta}{Parameter vector \eqn{\boldsymbol \beta}.}
	\item{phi}{The dispersion \eqn{\varphi}.}
	\item{method}{Different methods for calculating density. 
	method="real" uses real density, 
	"saddle" uses saddlepoint approximation, 
	"finverse" uses Fourier inversion, 
	"mWtrans" uses modified W-transformation. 
	Default is "saddle".}
	\item{m}{Vector of means. Default is \eqn{\mbox{\bf m}}=NULL in which \eqn{\mbox{\bf m}} is
	calculated by \eqn{m_i=-\log (-\mbox{\bf X}_{i\cdot}^\top \boldsymbol \beta)}. If \eqn{\mbox{\bf m}}
	is not NULL, then \eqn{\mbox{\bf X}} and \eqn{\boldsymbol \beta} are not used. }
}
\value{
	A vector with the same dimension as y.
	The value of the density of \eqn{\text{TBE}_\infty} calculated by different methods. 
}
\details{
    The \eqn{\text{TBE}_\infty} distribution is the exponential dispersion 
	model generated by the Landau distribution and also known as a Tweedie 
	model with power infinity. Tweedie models are exponential dispersion 
	models possessing a power variance function (VF) of the form \eqn{V(m)=m^\gamma, 
	\gamma \in (-\infty ,0\rbrack \cup \lbrack 1,\infty )}, where \eqn{m} 
	is the mean. And \eqn{\text{TBE}_\infty} possesses an 
	exponential variance function \eqn{e^m}, which can be considered as 
	a limiting case of the power variance functions of Tweedie.
	
    The density of \eqn{\text{TBE}_\infty} is defined but cannot be written 
	in closed form, and hence evaluation is very difficult. Therefore, some 
	approximations are generally used.
}
\references{
	Bar-Lev, S.K. (2023).
	The Exponential Dispersion Model Generated by the Landau Distribution
	---A Comprehensive Review and Further Developments.
	\emph{Mathematics}
	\bold{11}(20), 4343.
	\doi{10.3390/math11204343}
	
   Dunn, P. K. and Smyth, G. K. (2008).
   Evaluation of Tweedie exponential dispersion model densities by Fourier inversion.
   \emph{Statistics and Computing}, 
   \bold{18}, 73--86.
   \doi{10.1007/s11222-007-9039-6}

   Dunn, P. K. and Smyth, G. K. (2018).
   \emph{Generalized Linear Models with Examples in R}.
   Springer: Berlin/Heidelberg, Germany.

   Dunn, P. K. and Smyth, G. K. (2022).
   Evaluation of Tweedie Exponential Family Models; R Package Version 2.3.5; 2022.
   Available online: https://cran.r-project.org/web/packages/tweedie/tweedie.pdf

	Jorgensen, B. (1997).
	\emph{The Theory of Dispersion Models}.
	Chapman and Hall, London.
	
	Sidi, Avram (1988).  
	A user-friendly extrapolation method for
	oscillatory infinite integrals.
	\emph{Mathematics of Computation}
	\bold{51}(183), 249--266.
	\doi{10.1090/S0025-5718-1988-0942153-5}
}
\examples{
y <- seq(10, 20, by = 0.1)
n <- length(y)
p <- 10
x <- matrix(runif(n * p), nrow = n, ncol = p)
X <- cbind(1, x)
beta <- rep(-0.01, p)
m0 <- rep(5,n)
phi <- 0.5
dTBEinf(y, X, beta, phi, method="real")
dTBEinf(y, X, beta, phi)
dTBEinf(y, X, beta, phi, method="finverse", m=m0)
dTBEinf(y, X, beta, phi, method="mWtrans")
} 
