\name{logLikTBEinf}
\alias{logLikTBEinf}
\title{log-likelihood of TBE infinity Distributions}
\description{The log-likelihood for \eqn{\text{TBE}_\infty}}
\usage{
	logLikTBEinf(y, X, beta, phi, m = NULL)
}
\arguments{
	\item{y}{Vector of responses \eqn{\mbox{\bf y}\in \mathbb{R}^{n}}.}
	\item{X}{The design matrix \eqn{\mbox{\bf X}\in \mathbb{R}^{n\times p}}.}
	\item{beta}{Parameter vector \eqn{\boldsymbol \beta}.}
	\item{phi}{The dispersion \eqn{\varphi}.}
	\item{m}{Vector of means. Default is \eqn{\mbox{\bf m}}=NULL in which \eqn{\mbox{\bf m}} is
	calculated by \eqn{m_i=-\log (-\mbox{\bf X}_{i\cdot}^\top \boldsymbol \beta)}. If \eqn{\mbox{\bf m}}
	is not NULL, then \eqn{\mbox{\bf X}} and \eqn{\boldsymbol \beta} are not used.}
}
\value{
	The log-likelihood for \eqn{\text{TBE}_\infty}. A scalar.
}
\details{
	\code{logLikTBEinf} calculates the log-likelihood. 
}
\references{
	Bar-Lev, S.K. (2023).
	The Exponential Dispersion Model Generated by the Landau Distribution
	---A Comprehensive Review and Further Developments.
	\emph{Mathematics}
	\bold{11}(20), 4343.
	\doi{10.3390/math11204343}
}
\examples{
y <- seq(10, 20, by = 0.1)
n <- length(y)
p <- 10
x <- matrix(runif(n * p), nrow = n, ncol = p)
X <- cbind(1, x)
beta <- rep(-0.01, p)
m0 <- rep(5,n)
phi <- 0.5
logLikTBEinf(y, X, beta, phi)
logLikTBEinf(y, X, beta, phi, m=m0)
}