\name{TBEinf.Dev}
\alias{TBEinf.Dev}
\title{Deviance of TBE infinity Distributions}
\description{The total deviance for \eqn{\text{TBE}_\infty}}
\usage{
	TBEinf.Dev(y, X, beta, m = NULL)
}
\arguments{
	\item{y}{Vector of responses \eqn{\mbox{\bf y}\in \mathbb{R}^{n}}.}
	\item{X}{The design matrix \eqn{\mbox{\bf X}\in \mathbb{R}^{n\times p}}.}
	\item{beta}{Parameter vector \eqn{\boldsymbol \beta}.}
	\item{m}{Vector of means. Default is \eqn{\mbox{\bf m}}=NULL in which \eqn{\mbox{\bf m}} is
	calculated by \eqn{m_i=-\log (-\mbox{\bf X}_{i\cdot}^\top \boldsymbol \beta)}. If \eqn{\mbox{\bf m}}
	is not NULL, then \eqn{\mbox{\bf X}} and \eqn{\boldsymbol \beta} are not used.}
}
\value{
	The total deviance for \eqn{\text{TBE}_\infty}. A scalar.
}
\details{
	\code{TBEinf.Dev} calculates the total deviance. 
}
\references{
	Bar-Lev, S.K. (2023).
	The Exponential Dispersion Model Generated by the Landau Distribution
	---A Comprehensive Review and Further Developments.
	\emph{Mathematics}
	\bold{11}(20), 4343.
	\doi{10.3390/math11204343}

	Dunn, P. K. and Smyth, G. K. (2018).
   \emph{Generalized Linear Models with Examples in R}.
   Springer: Berlin/Heidelberg, Germany.
}
\examples{
y <- seq(10, 20, by = 0.1)
n <- length(y)
p <- 10
x <- matrix(runif(n * p), nrow = n, ncol = p)
X <- cbind(1, x)
beta <- rep(-0.01, p)
m0 <- rep(5,n)
TBEinf.Dev(y, X, beta)
TBEinf.Dev(y, X, beta, m=m0)
}